ScreenOne
PrimerPantalla, viewModel y nave por main
