+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/MainActivity.kt
@@ -3,127 +3,51 @@ package ar.edu.unlam.mobile.scaffold
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material3.Button
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import androidx.navigation.navArgument
-import ar.edu.unlam.mobile.scaffold.ui.components.BottomBar
-import ar.edu.unlam.mobile.scaffold.ui.screens.HomeScreen
-import ar.edu.unlam.mobile.scaffold.ui.screens.SecondaryScreen
+import ar.edu.unlam.mobile.scaffold.ui.screens.ScreenOne
 import ar.edu.unlam.mobile.scaffold.ui.theme.MyApplicationTheme
 import dagger.hilt.android.AndroidEntryPoint
 
+
 @AndroidEntryPoint
+
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             MyApplicationTheme {
-
-                Surface(
-
-                ) {
-                    pantalla1()
-
-                }
-            }
-        }
+                // Define tu sistema de navegación y rutas
+                val navController = rememberNavController()
 
 
-    }
+                    NavHost(
+                        navController = navController,
+                        startDestination = "screenOne"
+                    ) {
+                        composable("screenOne") { navBackStackEntry ->
+                            val navController = rememberNavController()
+                            ScreenOne(navController = navController)
+                        }
 
-    @Composable
-    fun pantalla1(modifier: Modifier = Modifier) {
+                        composable("secondaryScreen") {
+                            SecondaryScreen()
+                        }
 
-        val controller = rememberNavController()
-        val backgroundImageRes = painterResource(id = R.drawable.fondo_estrellas)
-        val backgroundModifier = Modifier
-            .fillMaxSize()
-            .background(Color.White)
 
-        // Buttons
-        //val keyboardController = LocalSoftwareKeyboardController.current
-        val textState = remember { mutableStateOf("") }
-        val buttonModifier = Modifier
-            .padding(16.dp)
-
-        Box(
-            modifier = Modifier.fillMaxSize()
-        ) {
-
-            Image(
-                painter = backgroundImageRes,
-                contentDescription = null,
-                modifier = backgroundModifier,
-                contentScale = ContentScale.Crop
-            )
-
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(16.dp),
-                verticalArrangement = Arrangement.Bottom,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = "Bienvenido A La Trivia De Star Wars",
-                    modifier = Modifier
-                        .padding(bottom = 16.dp)
-                        .padding(8.dp)
-                        .size(200.dp)
-                        .padding(16.dp),
-                    color = Color.White,
-                    fontSize = 24.sp
-                )
-                Button(
-                    onClick = {
-                        controller.navigate("pantalla2")
-                    },
-                    modifier = buttonModifier
-                ) {
-                    Text(text = "Iniciar Juego")
-                }
-
-                Button(
-                    onClick = {
-                        System.exit(0)
-                    },
-                    modifier = buttonModifier
-                ) {
-                    Text(text = "Salir del Juego")
+                    }
                 }
             }
         }
     }
-}
+
+    private fun SecondaryScreen() {
+        // Implementa la pantalla secundaria aquí
+    }
+
+
 /*@AndroidEntryPoint
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
